e2e(#974): invert the shift-enter spec — it must SEND, not refuse
`issue816-shift-enter-speaks.spec.ts` pinned the ruling that #974
reverses: it asserted the refusal copy visible, the draft untouched and
nothing on the wire. All three are now wrong, so the spec follows the
product rather than being weakened around it — same class as the two
vitests, and the reason this is a rewrite and not a deletion is that the
real-browser proof is exactly what the issue asks for.

It now presses Shift+Enter and waits for the line in the SCROLLBACK.
That is stronger than the dispatch barrier composeSend uses: an empty
textarea only says the pump took the draft, and the visible outcome the
operator was denied is the message appearing. The two remaining
assertions are the parts jsdom cannot show — that the deleted copy is
really gone from the rendered surface, and that the composer is empty
rather than holding the UA's own line break, which is why
`preventDefault` stayed.

Renamed to match the issue that owns the behaviour; nothing outside the
file referenced the old name.
